+++ stopped.py
@@ -1,8 +1,3 @@
-from bs4 import BeautifulSoup
-from check_update import check_update
-from datetime import datetime
-from time import time
-from pymongo import MongoClient
 from telebot import types
 
 import configparser
@@ -17,8 +12,6 @@ config.sections()
 config.read('bot.conf')
 
 TOKEN = config['RASTREIOBOT']['TOKEN']
-int_check = int(config['RASTREIOBOT']['int_check'])
-LOG_INFO_FILE = config['RASTREIOBOT']['text_log']
 
 bot = telebot.TeleBot(TOKEN)
